add unit test for core logical functions
