+++ .github/workflows/publish.yml
@@ -54,6 +54,9 @@ jobs:
       - name: Install
         run: npm ci
 
+      - name: Run unit tests
+        run: npm run test
+
       - name: Typecheck
         run: npm run typecheck
